+++ CHANGELOG.md
@@ -4,6 +4,13 @@ This file summarises the most important changes that went live on each release
 of Concierge. Please check the Wiki entry on the release process to understand
 how this file is formatted and how the process works.
 
+## [1.1.0]
+### Fixed
+- Avantio: add House(Casa) property type
+
+### Added
+- THH integrations: quote, book, cancel and synchronisation
+
 ## [1.0.0]
 ### Fixed
 - Avantio: add handling of `check_in_too_near` response code